Added minimap. For it to work, some changes had to be made:
In the LevelFactory the position of the finish line is found.
Hud constructor now takes in the position of the finish line(used in
Mainscreen).
In the Hud the minimap is constructed.

In the loadingscreen I zeroed the "countDown" variable, to make the
loading faster.

In the B2dAssetManager I changed which tmx map to take in(for the
minimap to work the finish line has to be constructed in tiled program).